pyproject.toml
[project]
name = "hatena-bookmark-api"
version = "0.1.0"
description = "はてなブックマーク「後で読む」取得ツール"
authors = [
    {name = "User"}
]
dependencies = [
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[project.scripts]
hatena-bookmarks = "hatena_bookmark_api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []